requirements.txt
requests
beautifulsoup4
Pillow
pyTelegramBotAPI
flask
python-dotenv
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1